Add files via upload
A small project, where we can input information about pizzas - name, topping, size and quantity. The information gets saved on a file.txt and remain stored for next time when the program runs. Each size have a different cost and when we view the menu, depending on the size and quantity we get the total cost.